Add annotation for route in controller Check naming for classes after reordering

+++ OCsf4/src/Controller/QuizGenController.php
@@ -8,14 +8,29 @@ use App\Form\QuizType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+/**
+ * @Route("/quizgen")
+ */
 class QuizGenController extends Controller
 {
+    /**
+     * Lists all quizs => redirect to quizdis
+     *
+     * @Route("/", name="oc_quizgen_homepage")
+     * @Method("GET")
+     */
     public function indexAction()
 	{	
 		return $this->redirect($this->generateUrl('oc_quizdis_select'));
 	}
-	
+	/**
+	 * Add a quiz
+	 *
+	 * @Route("/add", name="oc_quizgen_add")
+	 * @Method("GET", "POST")
+	 */
 	public function addAction(Request $request)
 	{
 		$quiz = new Quiz();
@@ -34,7 +49,7 @@ class QuizGenController extends Controller
 			
 			$form->handleRequest($request);
 		
-			if ($form->isValid()) {
+			if ($form->isSubmitted() && $form->isValid()) {
 				$em = $this->getDoctrine()->getManager();
 				$em->persist($quiz);
 				
@@ -53,55 +68,57 @@ class QuizGenController extends Controller
 
 				return $this->redirect($this->generateUrl('oc_quizgen_view', array('id' => $quiz->getId())));
 			}
-		
 		}
 		// On passe la méthode createView() du formulaire à la vue
 		// afin qu'elle puisse afficher le formulaire toute seule
-		return $this->render('OCQuizgenBundle:Default:add.html.twig', array(
+		return $this->render('OCQuizgen\add.html.twig', array(
 			'form' => $form->createView(),
 		));
 	} 
-	
-	public function editAction($id, Request $request)
+	/**
+	 * Edit a quiz
+	 *
+	 * @Route("/edit/{id}", name="oc_quizgen_edit", requirements={
+     * "id": "\d+" }))
+	 * @Method("GET", "POST")
+	 */
+	public function editAction( Request $request, Quiz $quiz)
 	{
-	
-		$em = $this->getDoctrine()->getManager();
-		
-		$quiz = $em
-			->getRepository('OCQuizgenBundle:Quiz')
-			->find($id)
-		;
 		if (null === $quiz) {
-			throw new NotFoundHttpException("Le quiz d'id ".$id." n'existe pas.");
+			throw new NotFoundHttpException("Le quiz n'existe pas.");
 		}
-		
 		$form = $this->createForm(new QuizType(), $quiz, array('user' => $this->getUser()));
-		
-		if ($form->handleRequest($request)->isValid()) {
-			foreach ($quiz->getQCMs() as $QCM) {
-				$QCM->setQuiz($quiz);
-			}
-			$em->flush();
-			$request->getSession()->getFlashBag()->add('notice', 'Vous avez modifié un quiz avec succès.');
-			return $this->redirect($this->generateUrl('oc_quizgen_view', array('id' => $quiz->getId())));
+		if($request->isMethod('POST')) {
+		    $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                foreach ($quiz->getQCMs() as $QCM) {
+                    $QCM->setQuiz($quiz);
+                }
+                $em->flush();
+                $request->getSession()
+                    ->getFlashBag()
+                    ->add('notice', 'Vous avez modifié un quiz avec succès.');
+                return $this->redirect($this->generateUrl('oc_quizgen_view', array(
+                    'id' => $quiz->getId()
+                )));
+            }
 		}
-		return $this->render('OCQuizgenBundle:Default:edit.html.twig', array(
+		return $this->render('OCQuizgen\edit.html.twig', array(
 			'form' => $form->createView(),
 			'quiz' => $quiz
 		));
 	} 
-  
-	public function viewAction($id)
+	/**
+	 * View a quiz
+	 *
+	 * @Route("/view/{id}", name="oc_quizgen_view", requirements={
+	 * "id": "\d+" }))
+	 * @Method("GET")
+	 */
+	public function viewAction(Quiz $quiz)
 	{
-		 
-		$em = $this->getDoctrine()->getManager();
-		// On récupère le $id
-		$quiz = $em
-			->getRepository('OCQuizgenBundle:Quiz')
-			->find($id)
-		;
 		if (null === $quiz) {
-			throw new NotFoundHttpException("Le quiz d'id ".$id." n'existe pas.");
+			throw new NotFoundHttpException("Le quiz n'existe pas.");
 		}
 		
 		$idAuteur=$quiz->getAuthor();
@@ -112,29 +129,29 @@ class QuizGenController extends Controller
 			$userManager = $this->get('fos_user.user_manager');
 			$auteur = $userManager->findUserBy(array('id' => $idAuteur))->getUsername();
 		}
-		
-		
+				
 		if(($idAuteur != 0) && ($this->getUser() != null) && $idAuteur == $this->getUser()->getId())
-			return $this->render('OCQuizgenBundle:Default:view.html.twig', array(
+			return $this->render('OCQuizgen\view.html.twig', array(
 				'quiz'		=> $quiz,
 				'auteur'	=> $auteur
 			));
 		else if ($acces == 'public')
-			return $this->render('OCQuizgenBundle:Default:viewpublic.html.twig', array(
+			return $this->render('OCQuizgen\viewpublic.html.twig', array(
 				'quiz'		=> $quiz,
 				'auteur'	=> $auteur
 			));
 		else
-			throw new NotFoundHttpException("Vous n'avez pas accès au quiz d'id ".$id);
-		
+			throw new NotFoundHttpException("Vous n'avez pas accès au quiz ");
 	}
-	
-	public function menuAction($limit)
+	/* 
+	 * Seems unused 
+	 */
+	private function menuAction($limit)
 	{
 		$listQuizs = $this
 			->getDoctrine()
 			->getManager()
-			->getRepository('OCQuizgenBundle:Quiz')
+			->getRepository('Quiz')
 			->findBy(
 				array('acces' => 'public'),
 				array('id' => 'desc'),
@@ -142,17 +159,22 @@ class QuizGenController extends Controller
 				0
 			)
 		;
-		return $this->render('OCQuizgenBundle:Default:menu.html.twig', array(
+		return $this->render('OCQuizgen\menu.html.twig', array(
 		  'listQuizs' => $listQuizs
 		));
 	}
-	
+	/**
+	 * Lists all public quizs
+	 * TODO remove fos_user management 
+	 * @Route("/list", name="oc_quizgen_list")
+	 * @Method("GET")
+	 */
 	public function listAction()
 	{
 		$listQuizs = $this
 			->getDoctrine()
 			->getManager()
-			->getRepository('OCQuizgenBundle:Quiz')
+			->getRepository('Quiz::class')
 			->findBy(
 				array('acces'=>'public'),
 				array('id' => 'desc')
@@ -169,25 +191,30 @@ class QuizGenController extends Controller
 			}
 		}
 		
-		return $this->render('OCQuizgenBundle:Default:list.html.twig', array(
+		return $this->render('OCQuizgen\list.html.twig', array(
 		  'listQuizs' => $listQuizs,
 		  'listNoms' => $listNoms
 		));
 	}
-	
+	/**
+	 * Lists a user quizs
+	 * TODO remove fos_user management
+	 * @Route("/mylist", name="oc_quizgen_mylist")
+	 * @Method("GET")
+	 */
 	public function mylistAction()
 	{
 		if($this->getUser() != NULL) {
 			$listQuizs = $this
 				->getDoctrine()
 				->getManager()
-				->getRepository('OCQuizgenBundle:Quiz')
+				->getRepository('Quiz::class')
 				->findBy(
 					array('author'=> $this->getUser()->getId()),
 					array('id' => 'desc')
 				)
 			;
-			return $this->render('OCQuizgenBundle:Default:mylist.html.twig', array(
+			return $this->render('OCQuizgen\mylist.html.twig', array(
 			  'listQuizs' => $listQuizs,
 			  'auteur' => $this->getUser()->getUsername()
 			));
@@ -195,7 +222,13 @@ class QuizGenController extends Controller
 		else 
 			return $this->redirect($this->generateUrl('fos_user_registration_register'));
 	}
-	
+	/**
+	 * Delete a quiz
+	 *
+	 * @Route("/delete/{id}", name="oc_quizgen_delete", requirements={
+	 * "id": "\d+" }))
+	 * @Method({"GET","DELETE"})
+	 */
 	public function deleteAction(Quiz $quiz)
 	{
 		// On crée un formulaire vide, qui ne contiendra que le champ CSRF
@@ -206,7 +239,7 @@ class QuizGenController extends Controller
 		if ($request->getMethod() == 'POST') {
 			$form->bind($request);
 
-			if ($form->isValid()) { // Ici, isValid ne vérifie donc que le CSRF
+			if ($form->isSubmitted() && $form->isValid()) { // Ici, isValid ne vérifie donc que le CSRF
 				// On supprime l'article
 				$em = $this->getDoctrine()->getManager();
 				$em->remove($quiz);
@@ -220,43 +253,48 @@ class QuizGenController extends Controller
 			}
 		}
 		// Si la requête est en GET, on affiche une page de confirmation avant de supprimer
-		return $this->render('OCQuizgenBundle:Default:delete.html.twig', array(
+		return $this->render('OCQuizgen\delete.html.twig', array(
 		  'quiz' => $quiz,
 		  'form'    => $form->createView()
 		));
 	}
 	
-	
-	public function duplicateAction($id, Request $request)
+	/**
+	 * Duplicate a quiz
+	 *
+	 * @Route("/delete/{id}", name="oc_quizgen_duplicate", requirements={
+	 * "id": "\d+" }))
+	 * @Method({"GET","POST"})
+	 */
+	public function duplicateAction(Request $request, Quiz $oldQuiz)
 	{	
-		$nQuiz = new Quiz();
+		$newQuiz = new Quiz();
 		
 		$em = $this->getDoctrine()->getManager();
-		//aQuiz pour ancien quiz
-		$aQuiz = $em
-			->getRepository('OCQuizgenBundle:Quiz')
-			->find($id)
-		;
-		if (null === $aQuiz) {
-			throw new NotFoundHttpException("Le quiz d'id ".$id." n'existe pas.");
+
+		if (null === $oldQuiz) {
+			throw new NotFoundHttpException("Le quiz n'existe pas.");
 		}
 		
 		//$nQuiz->setQuiz(aQuiz->getQuiz());
-		$nQuiz = clone $aQuiz;
+		$newQuiz = clone $oldQuiz;
 		
-		$form = $this->createForm(new QuizType(), $nQuiz, array('user' => $this->getUser()));
-		
-		if ($form->handleRequest($request)->isValid()) {
-			$em->persist($nQuiz);
+		$form = $this->createForm(new QuizType(), $newQuiz, array('user' => $this->getUser()));
+		if ($request->getMethod() == 'POST') {
+		    $form->handleRequest($request);
+		    /** TODO check that object is deeply copied and not referenced **/
+		if ($form->isSubmitted() && $form->isValid()) {
+			$em->persist($newQuiz);
 			
-			foreach ($nQuiz->getQCMs() as $QCM) {
-				$QCM->setQuiz($nQuiz);
+			foreach ($newQuiz->getQCMs() as $QCM) {
+				$QCM->setQuiz($newQuiz);
 			}
 			$em->flush();
 			$request->getSession()->getFlashBag()->add('notice', 'Vous avez dupliqué un quiz avec succès');
-			return $this->redirect($this->generateUrl('oc_quizgen_view', array('id' => $nQuiz->getId())));
+			return $this->redirect($this->generateUrl('oc_quizgen_view', array('id' => $newQuiz->getId())));
 		}
-		return $this->render('OCQuizgenBundle:Default:duplicate.html.twig', array(
+		}
+		return $this->render('OCQuizgen\duplicate.html.twig', array(
 			'form' => $form->createView(),
 		));
 	}
